requirements.txt
# --- Core Flask Framework ---
Flask==2.3.3
flask-cors==4.0.0
werkzeug==2.3.7
python-dotenv==1.0.0

# --- MongoDB Database ---
pymongo==4.5.0
certifi==2023.7.22

# --- Azure OpenAI ---
openai==1.10.0
# Let pip resolve tiktoken automatically

# --- Communication ---
requests==2.31.0

# --- Compatible LangChain Versions ---
langchain-core==0.1.16
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.5
langsmith==0.0.83

# --- Basic Data Handling ---
json5==0.9.14
pyyaml==6.0.1

# --- Production Server ---
gunicorn==21.2.0